# Speed up Studio development via SWC and keeping built pages in memory for longer