--- a/apps/studio/.babelrc
+++ b/apps/studio/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["next/babel"],
-  "plugins": []
-}