--- a/apps/studio/next.config.js
+++ b/apps/studio/next.config.js
@@ -356,6 +356,10 @@ const nextConfig = {
 
     return config
   },
+  onDemandEntries: {
+    maxInactiveAge: 24 * 60 * 60 * 1000,
+    pagesBufferLength: 100,
+  }
 }
 
 // module.exports = withBundleAnalyzer(nextConfig)