--- a/apps/studio/tsconfig.json
+++ b/apps/studio/tsconfig.json
@@ -18,7 +18,8 @@
     "incremental": true,
     "paths": {
       "@ui/*": ["./../../packages/ui/src/*"] // handle ui package paths
-    }
+    },
+    "useDefineForClassFields": true
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx"],
   "exclude": ["node_modules"]